+++ satella/coding/__init__.py
@@ -7,11 +7,12 @@ from .concurrent import Monitor, RMonitor
 from .decorators import precondition, for_argument, PreconditionError
 from .fun_static import static_var
 from .recast_exceptions import rethrow_as, silence_excs
+from .metaclasses import metaclass_maker
 
 __all__ = [
     'Monitor', 'RMonitor', 'merge_dicts',
     'for_argument',
     'precondition', 'PreconditionError',
     'rethrow_as', 'silence_excs',
-    'static_var'
+    'static_var', 'metaclass_maker'
 ]